add time js file update book js link modules dir
