+++ index.html
@@ -1,72 +1,78 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="index.css" />
+    <title>Awesome Books</title>
+  </head>
 
-<head>
- <meta charset="UTF-8">
- <meta http-equiv="X-UA-Compatible" content="IE=edge">
- <meta name="viewport" content="width=device-width, initial-scale=1.0">
- <link rel="stylesheet" href="index.css" />
- <title>Awesome Books</title>
-</head>
+  <body>
+    <header id="header">
+      <nav class="navbar">
+        <a href="#" id="nav-header">Awesome Books</a>
+        <ul class="navbar-list">
+          <li class="test">
+            <a class="listBtn active" href="#book-list-section">List </a>
+          </li>
+          <li id="new">
+            <a class="listBtn" href="#book-add-section">Add New</a>
+          </li>
+          <li id="contact">
+            <a class="listBtn" href="#book-contact-section">Contact</a>
+          </li>
+        </ul>
+      </nav>
+    </header>
 
-<body>
+    <main>
+      <div id="dateWrapper">
+        <p id="date"></p>
+      </div>
 
- <header id="header">
-  <nav class="navbar">
-   <a href="#" id="nav-header">Awesome Books</a>
-   <ul class="navbar-list">
-    <li class="test"><a class="listBtn active" href="#book-list-section">List </a></li>
-    <li id="new"><a class="listBtn" href="#book-add-section">Add New</a></li>
-    <li id="contact"><a class="listBtn" href="#book-contact-section">Contact</a></li>
-   </ul>
-  </nav>
- </header>
+      <div class="flex-container">
+        <section class="section active" id="book-list-section">
+          <h1 class="pageTitle">All awesome books</h1>
+          <div id="book-list-container">
+            <section class="book-list"></section>
+          </div>
+        </section>
 
- <main>
-  <div id="dateWrapper">
-   <p id="date"></p>
-  </div>
+        <section class="section" id="book-add-section">
+          <h1 class="h-add">Add a new book</h1>
+          <form id="formSection">
+            <label for="title">
+              <input type="text" id="title" placeholder="Title" required />
+            </label>
+            <label for="author">
+              <input type="text" id="author" placeholder="Author" required />
+            </label>
+            <button class="btn-submit" type="submit" value="Add">Add</button>
+          </form>
+        </section>
 
-  <div class="flex-container">
-   <section class="section active" id="book-list-section">
-    <h1 class="pageTitle">All awesome books</h1>
-    <div id="book-list-container">
-     <section class="book-list"></section>
-    </div>
-   </section>
+        <section class="section" id="book-contact-section">
+          <h1>Contact Information</h1>
+          <div>
+            <p>
+              Do you have any questions or you just want to say "Hello"?<br />You
+              can reach out to me!
+            </p>
+          </div>
 
-   <section class="section" id="book-add-section">
-    <h1 class="h-add">Add a new book</h1>
-    <form id="formSection">
-     <label for="title">
-  <input type="text" id="title" placeholder="Title" required>
-  </label>
-     <label for="author">
-   <input type="text" id="author" placeholder="Author" required>
-  </label>
-     <button class="btn-submit" type="submit" value="Add"> Add</button>
-    </form>
-   </section>
+          <ul id="contact-list">
+            <li>My email: nicomutale@gmail.com</li>
+            <li>Address : Livingstone, Zambia</li>
+          </ul>
+        </section>
+      </div>
+    </main>
 
-   <section class="section" id="book-contact-section">
-    <h1>Contact Information</h1>
-    <div>
-     <p>Do you have any questions or you just want to say "Hello"?<br>You can reach out to me!</p>
-    </div>
+    <footer id="footer">
+      <p>Copyright &copy; <span>Nicholas Mutale</span></p>
+    </footer>
 
-    <ul id="contact-list">
-     <li>My email: nicomutale@gmail.com</li>
-     <li>Address : Livingstone, Zambia</li>
-    </ul>
-   </section>
-  </div>
- </main>
-
- <footer id="footer">
-  <p>Copyright &copy; <span>Nicholas Mutale</span></p>
- </footer>
-
- <script src="index.js"></script>
-</body>
-
+    <script type="module" src="index.js"></script>
+  </body>
 </html>
